scripts/pyproject.toml
[tool.poetry]
name = "streamforge-scripts"
version = "0.1.0"
description = "Utility scripts for StreamForge"
authors = ["Sriram Sivaraman"]

[tool.poetry.dependencies]
python = "^3.10"
minio = "^7.2.4"
polars = "^0.20.7"
openpyxl = "^3.1.2"
httpx = "^0.27"
xlsx2csv = "^0.8.6"
trino = "^0.330"
psycopg2-binary = "^2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
